Created map lookup table from tilesheet
-All deltaTime is now a double rather than float
-Created a lookup tile table for the map
-Rendered a map to the screen according to the array of numbers

+++ RPG-Game/Bullet.cpp
@@ -18,9 +18,9 @@ void Bullet::Initialize(const sf::Vector2f& position, const sf::Vector2f& target
 	direction = Math::NormalizeVector(target - position);
 }
 
-void Bullet::Update(float deltaTime)
+void Bullet::Update(double deltaTime)
 {
-	rectangleShape.setPosition(rectangleShape.getPosition() + direction * m_speed * deltaTime);
+	rectangleShape.setPosition(rectangleShape.getPosition() + direction * m_speed * (float)deltaTime);
 }
 
 void Bullet::Draw(sf::RenderWindow& window)
